Module Zarr

zarr Provides an Ocaml implementation of the Zarr version 3 storage format specification. It supports creation of arrays and groups as well as chunking arrays along any dimension. One can store a Zarr hierarchy in memory or on disk. Zarr also supports reading zarr hierarchies created using other implementations, as long as they are spec-compliant.

Consult the examples and limitations for more info.

References

The Zarr Version 3 specification.

Zarr community site.

Node

module Node : sig ... end

This module provides functionality for manipulating Zarr nodes.

Metadata

module ArrayMetadata : sig ... end

A module which contains functionality to work with a parsed JSON Zarr array metadata document.

module GroupMetadata : sig ... end

A module which contains functionality to work with a parsed JSON Zarr group metadata document.

Storage

module Storage : sig ... end

Codecs

module Codecs : sig ... end

An array has an associated list of codecs. Each codec specifies a bidirectional transform (an encode transform and a decode transform). This module contains building blocks for creating and working with a chain of codecs.

Indexing

module Indexing : sig ... end

A module housing functions for creating and manipulating indices and slices for working with Zarr arrays.

Examples

Create, read & write array.

open Zarr
open Zarr.Node
open Zarr.Codecs
open Zarr.Storage

let store =
  Result.get_ok @@ FilesystemStore.open_or_create "testdata.zarr" in
let group_node = Result.get_ok @@ GroupNode.of_path "/some/group" in
FilesystemStore.create_group store group_node;
let array_node = Result.get_ok @@ ArrayNode.(group_node / "name") in
FilesystemStore.create_array
  ~codecs:[`Transpose [|2; 0; 1|]; `Bytes BE; `Gzip L2]
  ~shape:[|100; 100; 50|]
  ~chunks:[|10; 15; 20|]
  Bigarray.Float32 
  Float.neg_infinity
  array_node
  store;
let slice = Owl_types.[|R [0; 20]; I 10; R []|] in
let x =
  Result.get_ok @@
  FilesystemStore.get_array store array_node slice Bigarray.Float32 in
let x' =
  Owl.Dense.Ndarray.Generic.map
    (fun _ -> Owl_stats_dist.uniform_rvs 0. 10.) x in
FilesystemStore.set_array store array_node slice x';

Using sharding codec.

let config =
  {chunk_shape = [|5; 3; 5|]
  ;codecs = [`Transpose [|2; 0; 1|]; `Bytes LE; `Gzip L5]
  ;index_codecs = [`Bytes BE; `Crc32c]
  ;index_location = Start} in
let shard_node = Result.get_ok @@ ArrayNode.(group_node / "another") in
FilesystemStore.create_array
  ~codecs:[`ShardingIndexed config]
  ~shape:[|100; 100; 50|]
  ~chunks:[|10; 15; 20|]
  Bigarray.Complex32
  Complex.zero
  shard_node
  store;

Explore a Zarr hierarchy.

Functions to query a zarr hierarchy are provided. These include listing all nodes, finding children of a group node, resizing an array, deleting nodes, obtaining metadata of a node, and more.

let a, g = FilesystemStore.find_all_nodes store in
FilesystemStore.reshape store array_node [|25; 32; 10|];
let meta =
  Result.get_ok @@ FilesystemStore.group_metadata store group_node in
GroupMetadata.show meta;
FilesystemStore.array_exists store shard_node;
let a, g = FilesystemStore.find_child_nodes store group_node in
FilesystemStore.erase_group_node store group_node;

Extension Points

This library also provides custom extensions not defined in the version 3 specification. These are tabulated below:

Extension Point

Details

Data Types

char, complex32, int (63-bit integer), nativeint

Limitations

Although this implementation tries to be spec compliant, it does come with a few limitations:

Ocaml does not have support for unsigned integers as array data types and thus this library cannot support reading values of datatypes uint32, uint64 and complex128.