Module Zarr.ArrayMetadata

A module which contains functionality to work with a parsed JSON Zarr array metadata document.

type t

A type representing a parsed array metadata document.

val create : ?sep:Zarr__.Extensions.separator -> ?codecs:Codecs.Chain.t -> ?dimension_names:string option list -> ?attributes:Yojson.Safe.t -> shape:int array -> ('a, 'b) Stdlib.Bigarray.kind -> 'a -> int array -> t

create ~shape kind fv cshp Creates a new array metadata document with shape shape, fill value fv, data type kind and chunk shape cshp.

val encode : t -> string

encode t returns a byte string representing a JSON Zarr array metadata.

val decode : string -> (t, [> Zarr__.Metadata.error ]) Stdlib.result

decode s decodes a bytes string s into a ArrayMetadata.t type, and returns an error if the decoding process fails.

val shape : t -> int array

shape t returns the shape of the zarr array represented by metadata type t.

val ndim : t -> int

ndim t returns the number of dimension in a Zarr array.

val chunk_shape : t -> int array

chunk_shape t returns the shape a chunk in this zarr array.

val data_type : t -> string

data_type t returns the data type as specified in the array metadata.

val is_valid_kind : t -> ('a, 'b) Stdlib.Bigarray.kind -> bool

is_valid_kind t kind checks if kind is a valid Bigarray kind that matches the data type of the zarr array represented by this metadata type.

val fillvalue_of_kind : t -> ('a, 'b) Stdlib.Bigarray.kind -> 'a

fillvalue_of_kind t kind returns the fill value of uninitialized chunks in this zarr array given kind. Raises Failure if the kind is not compatible with this array's fill value.

val attributes : t -> Yojson.Safe.t

attributes t Returns a Yojson type containing user attributes assigned to the zarr array represented by t.

val dimension_names : t -> string option list

dimension_name t returns a list of dimension names. If none are defined then an empty list is returned.

val codecs : t -> Codecs.Chain.t

codecs t Returns a type representing the chain of codecs applied when decoding/encoding a Zarr array chunk.

val grid_shape : t -> int array -> int array

grip_shape t returns the shape of the Zarr array's regular chunk grid, as defined in the Zarr V3 specification.

val index_coord_pair : t -> int array -> int array * int array

index_coord_pair t coord maps a coordinate of this Zarr array to a pair of chunk index and coordinate within that chunk.

val chunk_indices : t -> int array -> int array list

chunk_indices t shp returns a list of all chunk indices that would be contained in a zarr array of shape shp given the regular grid defined in array metadata t.

val chunk_key : t -> int array -> string

chunk_key t idx returns a key encoding of a the chunk index idx.

val update_attributes : t -> Yojson.Safe.t -> t

update_attributes t json returns a new metadata type with an updated attribute field containing contents in json

val update_shape : t -> int array -> t

update_shape t new_shp returns a new metadata type containing shape new_shp.

val equal : t -> t -> bool

equal a b returns true if a b are equal array metadata documents and false otherwise.

val of_yojson : Yojson.Safe.t -> (t, string) Stdlib.result

of_yojson json converts a Yojson.Safe.t object into a ArrayMetadata.t and returns an error message upon failure.

val to_yojson : t -> Yojson.Safe.t

to_yojson t serializes an array metadata type into a Yojson.Safe.t object.