Module Zarr.Codecs

An array has an associated list of codecs. Each codec specifies a bidirectional transform (an encode transform and a decode transform). This module contains building blocks for creating and working with a chain of codecs.

type arraytoarray = [

| `Transpose of int array

]

The type of array -> array codecs.

type compression_level =

| L0

| L1

| L2

| L3

| L4

| L5

| L6

| L7

| L8

| L9

A type representing valid Gzip codec compression levels.

type fixed_bytestobytes = [

| `Crc32c

]

A type representing bytes -> bytes codecs that produce fixed sized encoded strings.

type variable_bytestobytes = [

| `Gzip of compression_level

]

A type representing bytes -> bytes codecs that produce variable sized encoded strings.

type bytestobytes = [

| fixed_bytestobytes

| variable_bytestobytes

]

The type of bytes -> bytes codecs.

type endianness =

| Little

| Big

A type representing the configured endianness of an array.

type loc =

| Start

| End

A type representing the location of a shard's index array in an encoded byte string.

type arraytobytes = [

| `Bytes of endianness

| `ShardingIndexed of shard_config

]

The type of array -> bytes codecs.

and shard_config = {

chunk_shape : int array;

codecs : bytestobytes shard_chain;

index_codecs : fixed_bytestobytes shard_chain;

index_location : loc;

}

A type representing the Sharding indexed codec's configuration parameters.

and 'a shard_chain = {

a2a : arraytoarray list;

a2b : arraytobytes;

b2b : 'a list;

}

A type representing the chain of codecs used to encode/decode a shard's bytes and its index array.

type codec_chain = [ arraytoarray | arraytobytes | bytestobytes ] list

A type used to build a user-defined chain of codecs when creating a Zarr array.

type error = [

| `Extension of string

| `Gzip of Ezgzip.error

| `Transpose_order of int array * string

| `CodecChain of string

| `Sharding of int array * int array * string

]

The type of errors returned upon failure when an calling a function on a Chain type.

module Ndarray = Owl.Dense.Ndarray.Generic

module Chain : sig ... end

A module containing functions to encode/decode an array chunk using a predefined set of codecs.