Module Codecs.Chain

A module containing functions to encode/decode an array chunk using a predefined set of codecs.

type t

A type representing a valid chain of codecs for decoding/encoding a Zarr array chunk.

val create : int list -> codec list -> t

create s c returns a type representing a chain of codecs defined by chain c and chunk shape s.

raises Bytes_to_bytes_invariant
if c contains more than one bytes->bytes codec.

raises Invalid_transpose_order
if c contains a transpose codec with invalid order array.

raises Invalid_zstd_level
if c contains a Zstd codec whose compression level is invalid.

raises Invalid_sharding_chunk_shape
if c contains a shardingindexed codec with an incorrect inner chunk shape.

val encode : t -> 'a Ndarray.t -> string

encode t x computes the encoded byte string representation of array chunk x.

val decode : t -> 'a array_repr -> string -> 'a Ndarray.t

decode t repr x decodes the byte string x using codec chain t and decoded representation type repr.

val (=) : t -> t -> bool

x = y returns true if chain x is equal to chain y, and false otherwise.

val of_yojson : int list -> Yojson.Safe.t -> (t, string) Stdlib.result

of_yojson x returns a code chain of type t from its json object representation.

val to_yojson : t -> Yojson.Safe.t

to_yojson x returns a json object representation of codec chain x.