Module Zarr.GroupMetadata

A module which contains functionality to work with a parsed JSON Zarr group metadata document.

type t

A type representing a parsed group metadata document.

val default : t

Return a group metadata type with default values for all fields.

val encode : t -> string

encode t returns a byte string representing a JSON Zarr group metadata.

val decode : string -> (t, [> `Store_read of string ]) Stdlib.result

decode s decodes a bytes string s into a t type, and returns an error if the decoding process fails.

val update_attributes : t -> Yojson.Safe.t -> t

update_attributes t json returns a new metadata type with an updated attribute field containing contents in json.

val show : t -> string

show t pretty-prints the contents of the group metadata type t.

val attributes : t -> Yojson.Safe.t

attributes t Returns a Yojson type containing user attributes assigned to the zarr group represented by t.