Module Zarr.Metadata

This module provides functionality for manipulating a Zarr node's metadata JSON document.

The Zarr V3 specification defines two types of metadata documents: array and group metadata. Both types are stored under the key zarr.json within the prefix of a group or array.

exception Parse_error of string

raised when parsing a metadata JSON document fails.

module FillValue : sig ... end

module Array : sig ... end

A module which contains functionality to work with a parsed JSON Zarr array metadata document.

module Group : sig ... end

A module which contains functionality to work with a parsed JSON Zarr group metadata document.