Module Metadata.FillValue

type t =

| Char of char
(*
A single character string.
*)

| Bool of bool
(*
Must be a JSON boolean.
*)

| Int of Stdint.uint64
(*
Value must be a JSON number with no fractional or exponent part that is within the representable range of the corresponding integer data type.
*)

| Float of float
(*
Value representing a JSON float.
*)

| FloatBits of float
(*
A JSON string specifying a byte representation of the float a hexstring.
*)

| IntComplex of Stdlib.Complex.t
(*
A JSON 2-element array of integers representing a complex number.
*)

| FloatComplex of Stdlib.Complex.t
(*
A JSON 2-element array of floats representing a complex number.
*)

| FFComplex of Stdlib.Complex.t

| FBComplex of Stdlib.Complex.t

| BFComplex of Stdlib.Complex.t

| BBComplex of Stdlib.Complex.t
(*
Provides an element value to use for uninitialised portions of a Zarr array. The permitted values depend on the data type.
*)