Module Zarr.Node

This module provides functionality for manipulating Zarr nodes.

A Zarr V3 node is associated with either a group or an array. All nodes in a hierarchy have a name and a path. The root node does not have a name and is the empty string "". Except for the root node, each node in a hierarchy must have a name, which is a string of unicode code points. The following constraints apply to node names:

must not be the empty string ("").

must not include the character "/".

must not be a string composed only of period characters, e.g. "." or "..".

must not start with the reserved prefix "__".

type error = [

| `Node_invariant of string

]

The error type for operations on node types. It is returned by functions that create an array or group node type when one or more of a node's invariants are not satisfied as defined in the Zarr V3 specification.

module GroupNode : sig ... end

module ArrayNode : sig ... end