Module Node.ArrayNode

type t

The type of an array node.

val create : GroupNode.t -> string -> t

create p n returns an array node with parent p and name n.

raises Failure
if node invariants are not satisfied.

val (/) : GroupNode.t -> string -> t

The infix operator alias of ArrayNode.create

val of_path : string -> t

of_path s returns an array node from string s.

raises Failure
if node invariants are not satisfied.

val to_path : t -> string

to_path n returns array node n as a string path.

val name : t -> string

name n returns the name of array node n.

val parent : t -> GroupNode.t

parent n returns parent group node of n.

val (=) : t -> t -> bool

x = y returns true if nodes x and y are equal, and false otherwise.

val ancestors : t -> GroupNode.t list

ancestors n returns ancestor group nodes of n.

val is_parent : t -> GroupNode.t -> bool

is_parent n g returns true if group node g is the immediate parent of array node n and false otherwise.

val to_key : t -> string

to_key n converts a node's path to a key, as defined in the Zarr V3 specification.

val to_metakey : t -> string

to_prefix n returns the metadata key associated with node n, as defined in the Zarr V3 specification.

val show : t -> string

show n returns a string representation of a node type.

val pp : Stdlib.Format.formatter -> t -> unit

pp fmt t pretty prints a node type value.