Module Zarr.Storage

type error

The error type of supported storage backends.

module type S = sig ... end

The public interface of all supported stores.

module type STORE = sig ... end

The module interface that all supported stores must implement.

module type MAKER = sig ... end

module Make : MAKER

A functor for minting a new storage type as long as it's argument module implements the STORE interface.

module MemoryStore : sig ... end

An in-memory storage backend for Zarr V3 hierarchy.

module FilesystemStore : sig ... end

A local filesystem storage backend for a Zarr V3 hierarchy.