Module Storage.FilesystemStore

A local filesystem storage backend for a Zarr V3 hierarchy.

include S

type t

The storage type.

val create_group : ?metadata:Zarr__.Metadata.GroupMetadata.t -> t -> Node.t -> unit

create_group ~meta t node creates a group node in store t containing metadata meta. This is a no-op if a node node is already a member of this store.

val create_array : ?sep:Zarr__.Extensions.separator -> ?dimension_names:string option list -> ?attributes:Yojson.Safe.t -> ?codecs:Codecs.chain -> shape:int array -> chunks:int array -> ('a, 'b) Stdlib.Bigarray.kind -> 'a -> Node.t -> t -> (unit, [> Codecs.error ]) Stdlib.result

create_array ~sep ~dimension_names ~attributes ~codecs ~shape ~chunks kind fill node t creates an array node in store t where:

Separator sep is used in the array's chunk key encoding.

Dimension names dimension_names and user attributes attributes are included in it's metadata document.

A codec chain defined by codecs.

The array has shape shape and chunk shape chunks.

The array has data kind kind and fill value fv.

This operation can fail if the codec chain is not well defined.

val array_metadata : Node.t -> t -> (Zarr__.Metadata.ArrayMetadata.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

array_metadata node t returns the metadata of array node node. This operation returns an error if:

The node is not a member of store t.

if node node is a group node.

val group_metadata : Node.t -> t -> (Zarr__.Metadata.GroupMetadata.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

group_metadata node t returns the metadata of group node node. This operation returns an error if:

The node is not a member of store t.

if node node is an array node.

val find_child_nodes : t -> Node.t -> (Node.t list * Node.t list, [> Zarr__.Storage_intf.error ]) Stdlib.result

find_child_nodes t n returns a tuple of child nodes of group node n. The first element of the tuple is a list of array child nodes, and the second element a list of child group nodes. This operation can fail if:

Node n is not a member of store t.

Node n is an array node of store t.

val find_all_nodes : t -> Node.t list

find_all_nodes t returns a list of all nodes in store t. If the store has no nodes, an empty list is returned.

val erase_node : t -> Node.t -> unit

erase_node t n erases node n from store t. This function erases all child nodes if n is a group node. If node n is not a member of store t then this is a no-op.

val is_member : t -> Node.t -> bool

is_member t n returns true if node n is a member of store t and false otherwise.

val set_array : Node.t -> Owl_types.index array -> ('a, 'b) Zarr__.Storage_intf.Ndarray.t -> t -> (unit, [> Zarr__.Storage_intf.error ]) Stdlib.result

set_array n s x t writes n-dimensional array x to the slice s of array node n in store t. This operation fails if:

the ndarray x size does not equal slice s.

the kind of x is not compatible with node n's data type as described in its metadata document.

If there is a problem decoding/encoding node n chunks.

val get_array : Node.t -> Owl_types.index array -> ('a, 'b) Stdlib.Bigarray.kind -> t -> (('a, 'b) Zarr__.Storage_intf.Ndarray.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

get_array n s k t reads an n-dimensional array of size determined by slice s from array node n. This operation fails if:

If there is a problem decoding/encoding node n chunks.

kind k is not compatible with node n's data type as described in its metadata document.

The slice s is not a valid slice of array node n.

val reshape : t -> Node.t -> int array -> (unit, [> Zarr__.Storage_intf.error ]) Stdlib.result

reshape t n shape resizes array node n of store t into new size shape. If this operation fails, an error is returned. It can fail if:

Node n is not a valid array node.

If shape does not have the same dimensions as node n's shape.

val create : ?file_perm:Unix.file_perm -> string -> t

create ~file_perm path returns a new filesystem Zarr V3 store. This * operatioin fails if path is a directory that already exists.

val open_store : ?file_perm:Unix.file_perm -> string -> (t, error) Stdlib.result

open_store ~file_perm path returns an existing filesystem Zarr V3 store. * This operatioin fails if path is not a Zarr store path.

val open_or_create : ?file_perm:Unix.file_perm -> string -> (t, error) Stdlib.result

open_or_create ~file_perm path returns an existing filesystem store * and creates it if it does not exist at path path. See documentation * for open_store and create for more information.