Module Storage.Make

A functor for minting a new storage type as long as it's argument module implements the STORE interface.

Parameters

module Io : Types.IO

Signature

module Deferred = Io.Deferred

type t = Io.t

The storage type.

module Group : sig ... end

module Array : sig ... end

val hierarchy : t -> (Node.Array.t list * Node.Group.t list) Deferred.t

hierarchy t returns p where p is a pair of lists representing all nodes in store t. The first element of the pair is a list of all array nodes, and the second element is a list of all group nodes. This operation returns a pair of empty lists if store t is empty.

raises Parse_error
if any node has invalid node_type metadata.

val clear : t -> unit Deferred.t

clear t clears the store t by deleting all nodes. If the store is already empty, this is a no-op.