Module type Storage.MAKER

Parameters

module M : sig ... end

Signature

type t = M.t

The storage type.

val create_group : ?metadata:Zarr__.Metadata.GroupMetadata.t -> t -> Node.GroupNode.t -> unit

create_group ~meta t node creates a group node in store t containing metadata meta. This is a no-op if node is already a member of this store.

val create_array : ?sep:[< `Dot | `Slash Slash ] -> ?dimension_names:string option list -> ?attributes:Yojson.Safe.t -> ?codecs:Codecs.chain -> shape:int array -> chunks:int array -> ('a, 'b) Stdlib.Bigarray.kind -> 'a -> Node.ArrayNode.t -> t -> (unit, [> Codecs.error ]) Stdlib.result

create_array ~sep ~dimension_names ~attributes ~codecs ~shape ~chunks kind fill node t creates an array node in store t where:

Separator sep is used in the array's chunk key encoding.

Dimension names dimension_names and user attributes attributes are included in it's metadata document.

A codec chain defined by codecs.

The array has shape shape and chunk shape chunks.

The array has data kind kind and fill value fv.

This operation can fail if the codec chain is not well defined.

val array_metadata : Node.ArrayNode.t -> t -> (Zarr__.Metadata.ArrayMetadata.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

array_metadata node t returns the metadata of array node node. This operation returns an error if node is not a member of store t.

val group_metadata : Node.GroupNode.t -> t -> (Zarr__.Metadata.GroupMetadata.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

group_metadata node t returns the metadata of group node node. This operation returns an error if node is not a member of store t.

val find_child_nodes : t -> Node.GroupNode.t -> (Node.ArrayNode.t list * Node.GroupNode.t list, [> Zarr__.Storage_intf.error ]) Stdlib.result

find_child_nodes t n returns a tuple of child nodes of group node n. This operation can fail if n is not a member of store t.

val find_all_nodes : t -> Node.ArrayNode.t list * Node.GroupNode.t list

find_all_nodes t returns Some p where p is a pair of lists representing all nodes in store t. The first element of the pair is a list of all array nodes, and the second element is a list of all group nodes. If the store has no nodes, None is returned.

val erase_group_node : t -> Node.GroupNode.t -> unit

erase_group_node t n erases group node n from store t. This also erases all child nodes of n. If node n is not a member of store t then this is a no-op.

val erase_array_node : t -> Node.ArrayNode.t -> unit

erase_array_node t n erases group node n from store t. This also erases all child nodes of n. If node n is not a member of store t then this is a no-op.

val group_exists : t -> Node.GroupNode.t -> bool

group_exists t n returns true if group node n is a member of store t and false otherwise.

val array_exists : t -> Node.ArrayNode.t -> bool

array_exists t n returns true if array node n is a member of store t and false otherwise.

val set_array : Node.ArrayNode.t -> Owl_types.index array -> ('a, 'b) Zarr__.Storage_intf.Ndarray.t -> t -> (unit, [> Zarr__.Storage_intf.error ]) Stdlib.result

set_array n s x t writes n-dimensional array x to the slice s of array node n in store t. This operation fails if:

the ndarray x size does not equal slice s.

the kind of x is not compatible with node n's data type as described in its metadata document.

If there is a problem decoding/encoding node n chunks.

val get_array : Node.ArrayNode.t -> Owl_types.index array -> ('a, 'b) Stdlib.Bigarray.kind -> t -> (('a, 'b) Zarr__.Storage_intf.Ndarray.t, [> Zarr__.Storage_intf.error ]) Stdlib.result

get_array n s k t reads an n-dimensional array of size determined by slice s from array node n. This operation fails if:

If there is a problem decoding/encoding node n chunks.

kind k is not compatible with node n's data type as described in its metadata document.

The slice s is not a valid slice of array node n.

val reshape : t -> Node.ArrayNode.t -> int array -> (unit, [> Zarr__.Storage_intf.error ]) Stdlib.result

reshape t n shape resizes array node n of store t into new size shape. If this operation fails, an error is returned. It can fail if shape does not have the same dimensions as n's shape. If node n is not a member of store t then this is a no-op.