Module type Storage.STORE

The module interface that all supported stores must implement.

module Deferred : Types.Deferred

type t

The storage type.

module Group : sig ... end

module Array : sig ... end

val find_all : t -> (Node.Array.t list * Node.Group.t list) Deferred.t

find_all t returns p where p is a pair of lists representing all nodes in store t. The first element of the pair is a list of all array nodes, and the second element is a list of all group nodes. This operation returns a pair of empty lists if store t is empty.

raises Parse_error
if any node has invalid node_type metadata.

val erase_all : t -> unit Deferred.t

erase_all t clears the store t by deleting all nodes. If the store is already empty, this is a no-op.