Module Zarr.Util

module ExtPoint : sig ... end

The type representing a JSON extension point metadata configuration.

module ArrayMap : sig ... end

A finite map over integer array keys.

module Result_syntax : sig ... end

Result monad operator syntax.

module Indexing : sig ... end

A module housing functions for creating and manipulating indices and slices for working with Zarr arrays.

val get_name : Yojson.Safe.t -> string

get_name c returns the name value of a JSON metadata extension point configuration of the form {"name": value, "configuration": ...}, as defined in the Zarr V3 specification.

val prod : int array -> int

prod x returns the product of the elements of x.

val max : int array -> int

max x returns the maximum element of an integer array x.

val create_parent_dir : string -> Unix.file_perm -> unit

create_parent_dir f p creates all the parent directories of file name f if they don't exist given file permissions p.

val sanitize_dir : string -> string

sanitize_dir d Chops off any trailing '/' in directory path d.