Module Zarr

module Node : sig ... end

This module provides functionality for manipulating Zarr nodes.

module Indexing : sig ... end

A module housing functions for creating and manipulating indices and slices for working with Zarr arrays.

module ArrayMetadata : sig ... end

A module which contains functionality to work with a parsed JSON Zarr array metadata document.

module GroupMetadata : sig ... end

A module which contains functionality to work with a parsed JSON Zarr group metadata document.

module Storage : sig ... end

module Codecs : sig ... end