Module Zarr.Codecs

An array has an associated list of codecs. Each codec specifies a bidirectional transform (an encode transform and a decode transform). This module contains building blocks for creating and working with a chain of codecs.

type arraytoarray = [

| `Transpose of int array

]

The type of array -> array codecs.

type compression_level =

| L0

| L1

| L2

| L3

| L4

| L5

| L6

| L7

| L8

| L9

A type representing valid Gzip codec compression levels.

type fixed_bytestobytes = [

| `Crc32c

]

A type representing bytes -> bytes codecs that produce fixed sized encoded strings.

type variable_bytestobytes = [

| `Gzip of compression_level

]

A type representing bytes -> bytes codecs that produce variable sized encoded strings.

type bytestobytes = [

| fixed_bytestobytes

| variable_bytestobytes

]

The type of bytes -> bytes codecs.

type endianness =

| LE

| BE

A type representing the configured endianness of an array.

type loc =

| Start

| End

A type representing the location of a shard's index array in an encoded byte string.

type fixed_arraytobytes = [

| `Bytes of endianness

]

The type of array -> bytes codecs.

type variable_array_tobytes = [

| `ShardingIndexed of shard_config

]

and shard_config = {

chunk_shape : int array;

codecs : [ arraytoarray | fixed_arraytobytes | `ShardingIndexed of shard_config | bytestobytes ] list;

index_codecs : [ arraytoarray | fixed_arraytobytes | fixed_bytestobytes ] list;

index_location : loc;

}

A type representing the Sharding indexed codec's configuration parameters.

type array_tobytes = [

| fixed_arraytobytes

| variable_array_tobytes

]

The type of array -> bytes codecs.

type codec_chain = [ arraytoarray | array_tobytes | bytestobytes ] list

A type used to build a user-defined chain of codecs when creating a Zarr array.

type error = [

| `Extension of string

| `Gzip of Ezgzip.error

| `Transpose_order of int array * string

| `CodecChain of string

| `Sharding of int array * int array * string

]

The type of errors returned upon failure when an calling a function on a Chain type.

type partial_setter = ?append:bool -> (int * string) list -> unit

type ('a, 'b) array_repr = {

kind : ('a, 'b) Stdlib.Bigarray.kind;

shape : int array;

fill_value : 'a;

}

The type summarizing the decoded/encoded representation of a Zarr array or chunk.

module Chain : sig ... end

A module containing functions to encode/decode an array chunk using a predefined set of codecs.