Module Codecs.Chain

A module containing functions to encode/decode an array chunk using a predefined set of codecs.

type t

A type representing a valid chain of codecs for decoding/encoding a Zarr array chunk.

val create : ('a, 'b) Zarr__.Util.array_repr -> codec_chain -> (t, [> error ]) Stdlib.result

create r c returns a type representing a chain of codecs defined by chain c and decoded array representation type r.

val default : t

default returns the default codec chain that contains only the required codecs as defined in the Zarr Version 3 specification.

val encode : t -> ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t -> (string, [> error ]) Stdlib.result

encode t x computes the encoded byte string representation of array chunk x. Returns an error upon failure.

val decode : t -> ('a, 'b) Zarr__.Util.array_repr -> string -> (('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Genarray.t, [> error ]) Stdlib.result

decode t repr x decodes the byte string x using codec chain t and decoded representation type repr. Returns an error upon failure.

val (=) : t -> t -> bool

x = y returns true if chain x is equal to chain y, and false otherwise.

val of_yojson : Yojson.Safe.t -> (t, string) Stdlib.result

of_yojson x returns a code chain of type t from its json object representation.

val to_yojson : t -> Yojson.Safe.t

to_yojson x returns a json object representation of codec chain x.