Module Zarr.Node

This module provides functionality for manipulating Zarr nodes.

A Zarr V3 node is associated with either a group or an array. All nodes in a hierarchy have a name and a path. The root node does not have a name and is the empty string "". Except for the root node, each node in a hierarchy must have a name, which is a string of unicode code points. The following constraints apply to node names:

must not be the empty string ("").

must not include the character "/".

must not be a string composed only of period characters, e.g. "." or "..".

must not start with the reserved prefix "__".

exception Node_invariant

raised when a node's invariants are violated.

exception Cannot_rename_root

raised when attempting to rename a root node.

module Group : sig ... end

module Array : sig ... end