Module Zarr.Storage

exception Invalid_resize_shape

raised when resizing a Zarr array with an incorrect shape.

exception Invalid_data_type

raised when supplied data type is not the same as Zarr array's.

exception Invalid_array_slice

raised when requesting a view of a Zarr array with an incorrect slice.

exception Key_not_found of string

raised when a node's chunk key or metadata key is found in a store.

exception Not_a_filesystem_store of string

raised when opening a file that as if it was a Filesystem Zarr store.

module type STORE = sig ... end

The module interface that all supported stores must implement.

module Make (Io : Types.IO) : STORE with type t = Io.t and module Deferred = Io.Deferred

A functor for minting a new storage type as long as it's argument module implements the STORE interface.